+++ R/allClasses.R
@@ -6,8 +6,7 @@
 setClass(Class = 'SRM', representation = representation(
           meta = 'tbl_df',
           transitions = 'tbl_df',
-          rawChrom = 'tbl_df',
-          transformedChrom = 'tbl_df',
+          chroms = 'tbl_df',
           peaks = 'tbl_df',
           header = 'tbl_df',
           processingLog = 'list'