+++ R/peakAsymmetry-method.R
@@ -7,9 +7,9 @@ setMethod('peakAsymmetry', signature = 'SRM',
 
             for (i in 1:nrow(object@peaks)) {
             rawChrom <-
-                object@rawChrom %>% dplyr::filter(
+                object@chroms %>% dplyr::filter(
                   sampleID == object@peaks$sampleID[i] &
-                    index == object@peaks$index[i] &
+                    filter == object@peaks$filter[i] &
                     rt >= object@peaks$rtmin[i] &
                     rt <= object@peaks$rtmax[i]
                 )
@@ -20,9 +20,16 @@ setMethod('peakAsymmetry', signature = 'SRM',
               peak_front_10 <-
                 rawChrom$rt[rawChrom$rt < rtmax][which.min(abs(rawChrom$int[rawChrom$rt < rtmax] - max(rawChrom$int) * 0.1))]
 
+              if(length(peak_front_10) == 0){
+                peak_front_10 <- 0
+              }
 
               peak_tail_10 <-
                 rawChrom$rt[rawChrom$rt > rtmax][which.min(abs(rawChrom$int[rawChrom$rt > rtmax] - max(rawChrom$int) * 0.1))]
+
+              if(length(peak_tail_10) == 0){
+                peak_tail_10 <- 0
+              }
 
               A10 <- rtmax - peak_front_10
               B10 <- peak_tail_10 - rtmax